last refactor before moving base tests back to api
