+++ spray-client/src/it/scala/com/heroku/platform/api/client/spray/DomainSpec.scala
@@ -4,9 +4,9 @@ package com.heroku.platform.api.client.spray
 import com.heroku.platform.api._
 
 
-abstract class DomainSpec(aj: ApiRequestJson with ApiResponseJson) extends SprayApiSpec(aj) {
+abstract class DomainSpec(aj: ApiRequestJson with ApiResponseJson) extends ApiSpec(aj) {
 
-  val implicits:DomainRequestJson with DomainResponseJson  = aj
+  val implicits: DomainRequestJson with DomainResponseJson = aj
 
   import implicits._
 
@@ -17,7 +17,7 @@ abstract class DomainSpec(aj: ApiRequestJson with ApiResponseJson) extends Spray
       val domain = create(Domain.Create(app.id, "foo.bar.baz.com"))
       val domainList = listAll(Domain.List(app.id))
       domainList.contains(domain) must be(true)
-      val domainInfo =  info(Domain.Info(app.id, domain.id))
+      val domainInfo = info(Domain.Info(app.id, domain.id))
       domainInfo must equal(domain)
       delete(Domain.Delete(app.id, domain.id))
     }
@@ -27,7 +27,7 @@ abstract class DomainSpec(aj: ApiRequestJson with ApiResponseJson) extends Spray
 }
 
 
-class SprayDomainSpec extends DomainSpec(SprayJsonBoilerplate)
+class SprayDomainSpec extends DomainSpec(SprayJsonBoilerplate) with SprayApiSpec
 
 
-class PlayDomainSpec extends DomainSpec(PlayJsonBoilerplate)
+class PlayDomainSpec extends DomainSpec(PlayJsonBoilerplate) with SprayApiSpec
